[INTERPRETER] * NEW: A new interpreter API that allows a component to tell that its   custom stream data is immediately available. * BUG: GB.Stream.SetBytesRead() now increments the internal number of read   bytes, so that the 'Line Input' buffer is taken into account.
[GB.NET.CURL]
* NEW: The FtpClient or HttpClient Read event is now called as long as 
  there is something to read *and* the event handler actually read at least
  one byte.
* BUG: The handle of a curl object is now -1 (i.e. it is not available).
* BUG: The data received by a curl object is immediately available to the
  READ instruction.


git-svn-id: svn://localhost/gambas/trunk@5553 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.net.curl/src/CFtpClient.c
@@ -74,7 +74,7 @@ static int ftp_write_curl(void *buffer, size_t size, size_t nmemb, void *_object
 	if (THIS->async)
 	{
 		GB.Ref(THIS);
-		GB.Post(CURL_raise_read,(long)THIS);
+		GB.Post(CURL_raise_read, (intptr_t)THIS);
 	}
 	
 	return nmemb;
